+++ constants.ts
@@ -1,5 +1,5 @@
 
-export const APP_VERSION = '4.1.3';
+export const APP_VERSION = '4.1.4';
 export const APP_MONTH = 'Mag 2026';
 
 export const INITIAL_ELO = 1500;